+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The version format is MCMAJOR.MCMINOR.MCPATCH-MAJOR.MINOR.PATCH, eg, 1.18.1-1.2.
 - Added configuration gui api
 - Added `IngredientStack` to support recipes which require stacked items
 - Added `IntContainerData`
+- Added fluid ingredients for use in custom recipes.
 
 ## [1.18.2-1.0.2a]
 ### Fixed